requirements.txt
# Web logger dashboard
flask
psutil
requests

# ArxivCast (intel-stack) - required when using /intel Search & Populate or podcast
arxiv
openai
edge-tts
pydub

# Voice conversation (STT, TTS) - optional; install for /intel voice chat
faster-whisper
piper-tts
scipy
